+++ CHANGELOG.md
@@ -51,6 +51,10 @@ This driver uses semantic versioning:
 
   This option was added in ArangoDB 3.11.
 
+- Added support for `allowRetry` option in `db.query`
+
+  This feature was added in ArangoDB 3.11.
+
 - Added `x-arango-driver` header
 
   The arangojs driver now correctly identifies itself to ArangoDB, allowing the